Begin component refactoring
