+++ front/src/modules/ui/layout/navbar/NavItemsContainer.tsx
@@ -1,0 +1,17 @@
+import styled from '@emotion/styled';
+
+type OwnProps = {
+  children: React.ReactNode;
+};
+
+const StyledNavItemsContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin-top: 40px;
+`;
+
+function NavItemsContainer({ children }: OwnProps) {
+  return <StyledNavItemsContainer>{children}</StyledNavItemsContainer>;
+}
+
+export default NavItemsContainer;
